backend/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
APScheduler>=3.10.0
requests>=2.31.0
